core/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
amqp==5.0.6
    # via kombu
asgiref==3.4.1
    # via django
billiard==3.6.4.0
    # via celery
celery==5.1.2
    # via
    #   -r requirements.in
    #   django-celery-results
certifi==2021.10.8
    # via requests
cffi==1.14.6
    # via cryptography
charset-normalizer==2.0.7
    # via requests
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
cryptography==35.0.0
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
django==3.2.8
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-filter
    #   django-modeltranslation
    #   djangorestframework
django-allauth==0.45.0
    # via -r requirements.in
django-celery-results==2.2.0
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
django-modeltranslation==0.17.3
    # via -r requirements.in
django-simple-history==3.0.0
    # via -r requirements.in
django-webpack-loader==0.7.0
    # via -r requirements.in
django-widget-tweaks==1.4.8
    # via -r requirements.in
djangorestframework==3.12.4
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.3
    # via requests
kombu==5.1.0
    # via celery
markdown==3.3.4
    # via -r requirements.in
oauthlib==3.1.1
    # via requests-oauthlib
pillow==8.4.0
    # via -r requirements.in
prompt-toolkit==3.0.20
    # via click-repl
psycopg2==2.9.1
    # via -r requirements.in
pycparser==2.20
    # via cffi
pyjwt[crypto]==2.2.0
    # via django-allauth
python3-openid==3.2.0
    # via django-allauth
pytz==2021.3
    # via
    #   celery
    #   django
requests==2.26.0
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.0
    # via django-allauth
six==1.16.0
    # via
    #   click-repl
    #   django-modeltranslation
sqlparse==0.4.2
    # via django
urllib3==1.26.7
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
